server/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
numpy
opencv-python
matplotlib
GitPython
tensorboard
cloudinary
